Move the variable extraction code to different file
